requirements.txt
langgraph>=0.2.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
chainlit>=1.3.0
pydantic>=2.0.0
python-dotenv>=1.0.0
langsmith>=0.1.0
openai>=1.0.0
